refactor: move `ApplicationCommandInteractionDataResolved` to base module
