+++ disnake/interactions/base.py
@@ -8,8 +8,16 @@ from typing import TYPE_CHECKING, Any, Dict, List, Mapping, Optional, Tuple, Uni
 
 from .. import utils
 from ..app_commands import OptionChoice
-from ..channel import ChannelType, PartialMessageable
-from ..enums import InteractionResponseType, InteractionType, Locale, WebhookType, try_enum
+from ..channel import PartialMessageable, _threaded_guild_channel_factory
+from ..enums import (
+    ChannelType,
+    InteractionResponseType,
+    InteractionType,
+    Locale,
+    OptionType,
+    WebhookType,
+    try_enum,
+)
 from ..errors import (
     HTTPException,
     InteractionNotEditable,
@@ -26,6 +34,7 @@ from ..member import Member
 from ..message import Attachment, Message
 from ..object import Object
 from ..permissions import Permissions
+from ..role import Role
 from ..ui.action_row import components_to_dict
 from ..user import ClientUser, User
 from ..webhook.async_ import Webhook, async_context, handle_message_parameters
@@ -34,6 +43,7 @@ __all__ = (
     "Interaction",
     "InteractionMessage",
     "InteractionResponse",
+    "InteractionDataResolved",
 )
 
 if TYPE_CHECKING:
@@ -41,13 +51,13 @@ if TYPE_CHECKING:
 
     from aiohttp import ClientSession
 
+    from ..abc import MessageableChannel
     from ..app_commands import Choices
-    from ..channel import CategoryChannel, StageChannel, TextChannel, VoiceChannel
     from ..client import Client
     from ..embeds import Embed
     from ..ext.commands import AutoShardedBot, Bot
     from ..file import File
-    from ..guild import GuildMessageable
+    from ..guild import GuildChannel, GuildMessageable
     from ..mentions import AllowedMentions
     from ..state import ConnectionState
     from ..threads import Thread
@@ -55,6 +65,7 @@ if TYPE_CHECKING:
     from ..types.interactions import (
         ApplicationCommandOptionChoice as ApplicationCommandOptionChoicePayload,
         Interaction as InteractionPayload,
+        InteractionDataResolved as InteractionDataResolvedPayload,
     )
     from ..ui.action_row import Components, MessageUIComponent, ModalUIComponent
     from ..ui.modal import Modal
@@ -62,14 +73,7 @@ if TYPE_CHECKING:
     from .message import MessageInteraction
     from .modal import ModalInteraction
 
-    InteractionChannel = Union[
-        VoiceChannel,
-        StageChannel,
-        TextChannel,
-        CategoryChannel,
-        Thread,
-        PartialMessageable,
-    ]
+    InteractionChannel = Union[GuildChannel, Thread, PartialMessageable]
 
     AnyBot = Union[Bot, AutoShardedBot]
 
@@ -1604,3 +1608,185 @@ class InteractionMessage(Message):
             asyncio.create_task(inner_call())
         else:
             await self._state._interaction.delete_original_response()
+
+
+class InteractionDataResolved(Dict[str, Any]):
+    """Represents the resolved data related to an interaction with an application command.
+
+    .. versionadded:: 2.1
+
+    .. versionchanged:: 2.7
+        Renamed from ``ApplicationCommandInteractionDataResolved`` to ``InteractionDataResolved``.
+
+    Attributes
+    ----------
+    members: Dict[:class:`int`, :class:`Member`]
+        A mapping of IDs to partial members (``deaf`` and ``mute`` attributes are missing).
+    users: Dict[:class:`int`, :class:`User`]
+        A mapping of IDs to users.
+    roles: Dict[:class:`int`, :class:`Role`]
+        A mapping of IDs to roles.
+    channels: Dict[:class:`int`, Channel]
+        A mapping of IDs to partial channels (only ``id``, ``name`` and ``permissions`` are included,
+        threads also have ``thread_metadata`` and ``parent_id``).
+    messages: Dict[:class:`int`, :class:`Message`]
+        A mapping of IDs to messages.
+    attachments: Dict[:class:`int`, :class:`Attachment`]
+        A mapping of IDs to attachments.
+
+        .. versionadded:: 2.4
+    """
+
+    __slots__ = ("members", "users", "roles", "channels", "messages", "attachments")
+
+    def __init__(
+        self,
+        *,
+        data: InteractionDataResolvedPayload,
+        state: ConnectionState,
+        guild_id: Optional[int],
+    ):
+        data = data or {}
+        super().__init__(data)
+
+        self.members: Dict[int, Member] = {}
+        self.users: Dict[int, User] = {}
+        self.roles: Dict[int, Role] = {}
+        self.channels: Dict[int, InteractionChannel] = {}
+        self.messages: Dict[int, Message] = {}
+        self.attachments: Dict[int, Attachment] = {}
+
+        users = data.get("users", {})
+        members = data.get("members", {})
+        roles = data.get("roles", {})
+        channels = data.get("channels", {})
+        messages = data.get("messages", {})
+        attachments = data.get("attachments", {})
+
+        guild: Optional[Guild] = None
+        # `guild_fallback` is only used in guild contexts, so this `MISSING` value should never be used.
+        # We need to define it anyway to satisfy the typechecker.
+        guild_fallback: Union[Guild, Object] = MISSING
+        if guild_id is not None:
+            guild = state._get_guild(guild_id)
+            guild_fallback = guild or Object(id=guild_id)
+
+        for str_id, user in users.items():
+            user_id = int(str_id)
+            member = members.get(str_id)
+            if member is not None:
+                self.members[user_id] = (
+                    guild
+                    and guild.get_member(user_id)
+                    or Member(
+                        data=member,
+                        user_data=user,
+                        guild=guild_fallback,  # type: ignore
+                        state=state,
+                    )
+                )
+            else:
+                self.users[user_id] = User(state=state, data=user)
+
+        for str_id, role in roles.items():
+            self.roles[int(str_id)] = Role(
+                guild=guild_fallback,  # type: ignore
+                state=state,
+                data=role,
+            )
+
+        for str_id, channel in channels.items():
+            channel_id = int(str_id)
+            factory, _ = _threaded_guild_channel_factory(channel["type"])
+            if factory:
+                channel["position"] = 0  # type: ignore
+                self.channels[channel_id] = (
+                    guild
+                    and guild.get_channel(channel_id)
+                    or factory(
+                        guild=guild_fallback,  # type: ignore
+                        state=state,
+                        data=channel,  # type: ignore
+                    )
+                )
+            else:
+                self.channels[channel_id] = PartialMessageable(
+                    state=state, id=channel_id, type=try_enum(ChannelType, channel["type"])
+                )
+
+        for str_id, message in messages.items():
+            channel_id = int(message["channel_id"])
+            channel = cast(
+                "Optional[MessageableChannel]",
+                (guild and guild.get_channel(channel_id) or state.get_channel(channel_id)),
+            )
+            if channel is None:
+                # The channel is not part of `resolved.channels`,
+                # so we need to fall back to partials here.
+                channel = PartialMessageable(state=state, id=channel_id, type=None)
+            self.messages[int(str_id)] = Message(state=state, channel=channel, data=message)
+
+        for str_id, attachment in attachments.items():
+            self.attachments[int(str_id)] = Attachment(data=attachment, state=state)
+
+    def __repr__(self):
+        return (
+            f"<InteractionDataResolved members={self.members!r} users={self.users!r} "
+            f"roles={self.roles!r} channels={self.channels!r} messages={self.messages!r} attachments={self.attachments!r}>"
+        )
+
+    def get_with_type(self, key: Any, option_type: OptionType, default: Any = None):
+        if isinstance(option_type, int):
+            option_type = try_enum(OptionType, option_type)
+        if option_type is OptionType.mentionable:
+            key = int(key)
+            result = self.members.get(key)
+            if result is not None:
+                return result
+            result = self.users.get(key)
+            if result is not None:
+                return result
+            return self.roles.get(key, default)
+
+        if option_type is OptionType.user:
+            key = int(key)
+            member = self.members.get(key)
+            if member is not None:
+                return member
+            return self.users.get(key, default)
+
+        if option_type is OptionType.channel:
+            return self.channels.get(int(key), default)
+
+        if option_type is OptionType.role:
+            return self.roles.get(int(key), default)
+
+        if option_type is OptionType.attachment:
+            return self.attachments.get(int(key), default)
+
+        return default
+
+    def get(self, key: int):
+        if key is None:
+            return None
+
+        res = self.members.get(key)
+        if res is not None:
+            return res
+        res = self.users.get(key)
+        if res is not None:
+            return res
+        res = self.roles.get(key)
+        if res is not None:
+            return res
+        res = self.channels.get(key)
+        if res is not None:
+            return res
+        res = self.messages.get(key)
+        if res is not None:
+            return res
+        res = self.attachments.get(key)
+        if res is not None:
+            return res
+
+        return None
