+++ docs/api.rst
@@ -5098,6 +5098,14 @@ InteractionMessage
     :members:
     :inherited-members:
 
+InteractionDataResolved
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+.. attributetable:: InteractionDataResolved
+
+.. autoclass:: InteractionDataResolved()
+    :members:
+
 ApplicationCommandInteractionData
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
@@ -5112,14 +5120,6 @@ ApplicationCommandInteractionDataOption
 .. attributetable:: ApplicationCommandInteractionDataOption
 
 .. autoclass:: ApplicationCommandInteractionDataOption()
-    :members:
-
-ApplicationCommandInteractionDataResolved
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. attributetable:: ApplicationCommandInteractionDataResolved
-
-.. autoclass:: ApplicationCommandInteractionDataResolved()
     :members:
 
 MessageInteractionData
